Rename the random file name back to SearchBar
